requirements.txt
# Edoras Flask API - Python Dependencies

# Web Framework
Flask==2.3.3
flask-cors==4.0.0

# Database
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
alembic==1.12.0

# Authentication
Flask-JWT-Extended==4.5.3

# Serialization & Validation
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0
marshmallow==3.20.1

# Environment & Configuration
python-dotenv==1.0.0

# Azure SDK
azure-storage-blob==12.17.0
azure-keyvault-secrets==4.7.0
azure-identity==1.14.0

# Database Drivers
# Per Azure SQL Server
pyodbc==4.0.39
# Per PostgreSQL (alternativa)
psycopg2-binary==2.9.7
# Per MySQL (alternativa)
PyMySQL==1.1.0

# Utilities
requests==2.31.0
Werkzeug==2.3.7

# Production Server
gunicorn==21.2.0

# Azure App Service optimizations
whitenoise==6.5.0
